+++ src/App.css
@@ -34,7 +34,7 @@
   color: #61dafb;
 }
 
-.formWrap {
+.form {
   display: flex;
   flex-wrap: wrap;
   justify-content: center;
@@ -44,15 +44,20 @@
   /* color: #303030;  set in App.js */
 }
 
-.formElementWrap {
+.form-group {
   display: flex;
   height: 1.2em;
   line-height: 1.2em;
   justify-content: center;
   padding-bottom: 0.25em;
 }
-.formElementWrap input {
+.form-group input {
   margin: 0em 0.5em;
+  padding-left: 5px;
+}
+.form-group button {
+  padding: 0 10px;
+  letter-spacing: 0.75px;
 }
 
 .resetImg {
@@ -125,6 +130,7 @@ tr {
 }
 .pointCol {
   text-align: center;
+  font-family: monospace;
 }
 .authorCol {
   font-size: 0.85em;
@@ -132,6 +138,7 @@ tr {
 .dateCol {
   min-width: 6em !important;
   font-size: 0.9em;
+  font-family: monospace;
 }
 
 .linkSeparator {
