+++ src/components/index.js
@@ -1,2 +1,3 @@
 export * from './DataTable';
 export * from './Form';
+export * from './Header';
